# [spring-webflux] Leverage R2BDC recent optimizations